add initial service worker
This service worker will mock the remote API for saving, loading, and
otherwise managing playlists and other data. This enables code
deduplication as the code for saving locally and saving remotely is
the same, except for the API URL.

Service workers use import statements in SvelteKit[0] when running in
dev mode. Firefox does not support this[1], so for development, either
a Chromium-based browser or Safari is now necessary. Hope this gets
implemented soon in Firefox.

[0] https://svelte.dev/docs/kit/service-workers#During-development
[1] https://caniuse.com/mdn-api_serviceworker_ecmascript_modules